+++ src/services/video-library/videoLibrary.js
@@ -29,6 +29,7 @@ const { getRelativeMediaPath, getAbsoluteMediaPath } = require('../service-utils
 const { resetMetaData } = require('../database/metaDataDbService');
 const responseStatus = require('../../constants/responseStatus');
 const mediaTypes = require('../../constants/mediaTypes');
+const { addMediaLocation } = require('../location/location');
 
 const onImportFileSuccess = (data) => {
     if (data.mediaType !== mediaTypes.VIDEO) return;
@@ -43,6 +44,9 @@ const onImportFileSuccess = (data) => {
     const doesVideoExist = checkForDuplicate(videoStats.id);
     if (!doesVideoExist) {
         addVideo(videoStats);
+        if (videoStats.latitude && videoStats.longitude) {
+            addMediaLocation(mediaTypes.VIDEO, videoStats.id, videoStats.latitude, videoStats.longitude);
+        }
     }
 };
 
@@ -211,7 +215,7 @@ const importVideoFromWatchedDirectory = async (videoDetails, destinationCollecti
 
             serviceEventBus.publish(interServiceEvents.IMPORT_FILE_SUCCESS, {
                 completedMediaStats: newMediaStats,
-                mediaType: 'video',
+                mediaType: mediaTypes.VIDEO,
             });
 
             if (userPreferenceStore.get('showVideoPreviewOnHover')) {